f2c: Changed return type of r_sign(real *, real *) from double to float

+++ SRC/csteqr.c
@@ -31,7 +31,7 @@ static real c_b41 = 1.f;
 
     /* Builtin functions */
     float sqrtf(real);
-    double r_sign(real *, real *);
+    float r_sign(real *, real *);
 
     /* Local variables */
     real b, c__, f, g;
